refactor: Move enum to model package
